add prioriti,status boring commit

+++ 1Jirafapi/src/main/java/scanpackage/service/TaskService.java
@@ -6,8 +6,11 @@ import java.util.List;
 
 public interface TaskService {
     List<TaskModel> getAll();
-    TaskModel findById(long id);
+
+    TaskModel findById(Integer id);
+
     TaskModel save(TaskModel task);
+
     TaskModel update(TaskModel task);
 
 }
